fix(legacy): update sector argument to match flash_erase_sector size

+++ legacy/supervise.c
@@ -45,7 +45,7 @@ static void svhandler_flash_program(uint32_t psize) {
   FLASH_CR |= FLASH_CR_PG;
 }
 
-static void svhandler_flash_erase_sector(uint16_t sector) {
+static void svhandler_flash_erase_sector(uint8_t sector) {
   /* we only allow erasing storage sectors 2 and 3. */
   if (sector < FLASH_STORAGE_SECTOR_FIRST ||
       sector > FLASH_STORAGE_SECTOR_LAST) {
